requirements.txt
blinker==1.7.0
cachelib==0.9.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.2
Flask-JWT-Extended==4.6.0
Flask-Testing==0.8.1
googlemaps==4.10.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
python-google-places==1.4.2
requests==2.31.0
secure-smtplib==0.1.1
six==1.16.0
urllib3==2.2.1
Werkzeug==3.0.1
